Make main functionality of RPS game
This commit is composed of the relatively basic features of a rock,
paper, scissor, shoot game. It runs in the console and only has a few
quality of life features to make it readable for players.

It employs several different functions in a smooth and what I believe is
an efficient method a couple of the tricks I've picked up so far.