Add UI to existing RPS game
I added several DOM items which allows this game to be played outside of
the terminal. Three of which are buttons and are taken as inputs, the
rest are DIVs which contain the responsive outputs for the user. There
are also some safety features in place which make sure that the only
inputs accepted are through the buttons.

This is a finished version of the UI and can be safely published to the
main branch.